refactor: move chain service v0 into the chain-service directory
