+++ packages/server-wallet/src/chain-service/chain-service-v0/assets/onchain-architecture.png
